Make some contextual analyzer cell analyzers
